API/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
SQLAlchemy==1.4.27
PyMySQL==1.0.2
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.4.0
python-multipart==0.0.20
pydantic==1.8.2
starlette==0.37.2
cryptography==36.0.0   
requests==2.32.3       
PyJWT==2.10.1  